Gardening: Represent ServerFeature value as int
Motivation
----------
Shorts are awkward to work with, and require casts
when initializing with hex literals.

Modifications
-------------
Represent the value as an int instead of a short.
Assert the int value fits in 16 bits.

Add Javadoc for REPORT_UNIT_USAGE, copied from server.

Annotate the collections capability as having been
introduced in Couchbase 7.0 when it went GA,
instead of 6.5 when it was in developer preview.

Reviewed-on: https://review.couchbase.org/c/couchbase-jvm-clients/+/193069
Tested-by: Build Bot <build@couchbase.com>

+++ core-io/src/test/java/com/couchbase/client/core/io/netty/kv/ServerFeatureTest.java
@@ -27,7 +27,7 @@ class ServerFeatureTest {
     for (ServerFeature f : ServerFeature.values()) {
       assertEquals(f, ServerFeature.from(f.value()));
     }
-    short someUnknownValue = (short) 0xffff;
+    int someUnknownValue = 0xffff;
     assertNull(ServerFeature.from(someUnknownValue));
   }
 }
